Sort chapters correctly by date
The date listed on each entry will not be accurate (meaning it will not match the date that the chapter was actually published on AO3). It appears that `ao3-api` does not support pulling this information at this time.

Previously, no date was given at all, so clients may have assumed that the publish date was the current time. Some clients may also have shown the chapters in the wrong order as a result of this. 

This may not work correctly if all chapters were published on the same date. However, this is likely uncommon.